feat: Add CourseReq DTO, Booking and Review entities, and Booking repository
- Introduced CourseReq DTO for course creation requests.
- Created Booking entity with fields for order ID, tutor ID, student ID, date, hour, slot locked status, and booking status.
- Added Review entity to capture student feedback on courses, including rating and comments.
- Implemented BookingRepo interface for booking-related database operations, including methods for finding bookings by tutor and student IDs, and managing future bookings.
- Removed obsolete BookingRepository interface.
- Updated OrderService and PaymentService to reflect changes in course and user handling.
- Refactored TeacherCourseService and TutorProfileService to align with new entity structures.
- Deleted outdated SQL seed and schema files.

+++ src/main/java/com/learning/api/service/PaymentService.java
@@ -29,7 +29,7 @@ public class PaymentService {
         User user = userRepository.findById(order.getUserId()).orElse(null);
         if (user == null || user.getWallet() < totalCost) return false;// 餘額不足
 
-        user.setWallet((user.getWallet()-totalCost));
+        user.setWallet((int)(user.getWallet()-totalCost));
         userRepository.save(user);
 
         WalletLog log =new WalletLog();
